[DESIGN] feature/MBTI2 -> dev 변경사항 : activity_post_detail.xml, activity_post_item.xml, activity_post_list.xml, activity_post_my_detail.xml 디자인 변경
