+++ app/src/main/java/com/example/mbti_talk/MyProfile/MyProfileFragment.kt
@@ -8,7 +8,7 @@ import android.view.ViewGroup
 import androidx.appcompat.widget.AppCompatImageView
 import androidx.fragment.app.Fragment
 import com.bumptech.glide.Glide
-import com.example.mbti_talk.MBTI.SignUpMbtiActivity
+import com.example.mbti_talk.MBTI.MbtiActivity
 import com.example.mbti_talk.UserData
 import com.example.mbti_talk.databinding.FragmentMyProfileBinding
 import com.google.firebase.auth.FirebaseAuth
@@ -36,7 +36,7 @@ class MyProfileFragment : Fragment() {
         val view = binding.root
 
         binding.ProfileMbtiBtn.setOnClickListener {
-            val intent = Intent(requireContext(), SignUpMbtiActivity::class.java)
+            val intent = Intent(requireContext(), MbtiActivity::class.java)
             startActivity(intent)
         }
